+++ chartlets.js/packages/lib/src/plugins/mui/LinearProgress.tsx
@@ -21,9 +21,9 @@ export const LinearProgress = ({
   style,
   value,
   variant,
-  visible = true,
+  hidden = true,
 }: LinearProgressProps) => {
-  if (!visible) {
+  if (!hidden) {
     return null;
   }