+++ chartlets.py/chartlets/component.py
@@ -38,8 +38,8 @@ class Component(ABC):
     color: str | None = None
     """HTML `color` attribute. Optional."""
 
-    visible: bool | None = None
-    """If set, controls whether the component is visible."""
+    hidden: bool | None = None
+    """If set, controls whether the component is hidden."""
 
     children: list[Union["Component", str, None]] | None = None
     """Children used by many specific components. Optional."""